Perf: File structure
