+++ ui/src/enum/index.ts
@@ -1,0 +1,45 @@
+export enum FormatterMessageType {
+  PING = 'PING',
+  TERMINAL_INIT = 'TERMINAL_INIT',
+  TERMINAL_DATA = 'TERMINAL_DATA',
+  TERMINAL_SHARE = 'TERMINAL_SHARE',
+  TERMINAL_RESIZE = 'TERMINAL_RESIZE',
+  TERMINAL_K8S_DATA = 'TERMINAL_K8S_DATA',
+  TERMINAL_K8S_RESIZE = 'TERMINAL_K8S_RESIZE',
+  TERMINAL_SHARE_USER_REMOVE = 'TERMINAL_SHARE_USER_REMOVE',
+  TERMINAL_GET_SHARE_USER = 'TERMINAL_GET_SHARE_USER'
+}
+
+export enum SendLunaMessageType {
+  PING = 'PING',
+  CLOSE = 'CLOSE',
+  PONG = 'PONG',
+  CONNECT = 'CONNECT',
+  TERMINAL_ERROR = 'TERMINAL_ERROR',
+  MESSAGE_NOTIFY = 'MESSAGE_NOTIFY'
+}
+
+export enum ZmodemActionType {
+  ZMODEM_START = 'ZMODEM_START',
+  ZMODEM_END = 'ZMODEM_END'
+}
+
+export enum MessageType {
+  PING = 'PING',
+  CLOSE = 'CLOSE',
+  ERROR = 'ERROR',
+  CONNECT = 'CONNECT',
+  TERMINAL_SHARE = 'TERMINAL_SHARE',
+  TERMINAL_ERROR = 'TERMINAL_ERROR',
+  MESSAGE_NOTIFY = 'MESSAGE_NOTIFY',
+  TERMINAL_ACTION = 'TERMINAL_ACTION',
+  TERMINAL_SESSION = 'TERMINAL_SESSION',
+  TERMINAL_PERM_VALID = 'TERMINAL_PERM_VALID',
+  TERMINAL_SHARE_JOIN = 'TERMINAL_SHARE_JOIN',
+  TERMINAL_SHARE_LEAVE = 'TERMINAL_SHARE_LEAVE',
+  TERMINAL_PERM_EXPIRED = 'TERMINAL_PERM_EXPIRED',
+  TERMINAL_SESSION_PAUSE = 'TERMINAL_SESSION_PAUSE',
+  TERMINAL_SESSION_RESUME = 'TERMINAL_SESSION_RESUME',
+  TERMINAL_GET_SHARE_USER = 'TERMINAL_GET_SHARE_USER',
+  TERMINAL_SHARE_USER_REMOVE = 'TERMINAL_SHARE_USER_REMOVE'
+}
