+++ ui/src/types/modules/setting.type.ts
@@ -1,0 +1,19 @@
+import type { FunctionalComponent } from 'vue';
+
+export interface SettingConfig {
+  theme?: string;
+  drawerTitle: string;
+
+  items: Array<{
+    type: 'select' | 'button' | 'create' | 'list';
+    label: string;
+    labelIcon: FunctionalComponent;
+    labelStyle: {
+      fontSize: string;
+    };
+    showMore?: boolean;
+    disabled?: boolean;
+    value?: string;
+    options?: any;
+  }>;
+}
